devide logic
